# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CricZone API",
    "version": "1.0.0",
    "description": "Core authentication and cricket match endpoints. Additional controller coverage is tracked as a documentation follow-up.",
    "license": { "name": "ISC", "url": "https://opensource.org/license/isc-license-txt/" }
  },
  "servers": [
    {
      "url": "/",
      "description": "Current CricZone deployment"
    }
  ],
...
  "security": [],
  "paths": {
...
    "/api/matches": {
      "get": {
        "tags": ["Matches"],
        "summary": "List matches",
        "operationId": "getAllMatches",
        "parameters": [
          { "$ref": "#/components/parameters/Page" },
          { "$ref": "#/components/parameters/Limit" }
        ],
        "responses": {
          "200": {
            "description": "Paginated matches",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MatchListResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      },
      "post": {
        "tags": ["Matches"],
        "summary": "Create a match",
        "operationId": "createMatch",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateMatchInput" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Match created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/MatchResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "default": 1 }
      },
      "Limit": {
        "name": "limit",
        "in": "query",
        "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "Unauthorized": {
        "description": "Authentication required or invalid",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["success", "message"],
        "properties": {
          "success": { "type": "boolean", "example": false },
          "message": { "type": "string" },
          "requestId": { "type": "string" },
          "details": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        }
      },
...
      "PlayerInput": {
        "oneOf": [
          { "type": "string", "maxLength": 100 },
          {
            "type": "object",
            "properties": {
              "name": { "type": "string", "maxLength": 100 },
              "email": { "type": "string", "format": "email" },
              "userId": { "type": "string" }
            },
            "additionalProperties": true
          }
        ]
      },
      "CreateMatchInput": {
        "type": "object",
        "required": ["matchName", "venue", "matchDate"],
        "properties": {
          "matchName": { "type": "string", "maxLength": 120 },
          "matchType": { "type": "string", "enum": ["T20", "ODI", "Test", "Custom"], "default": "T20" },
          "customOvers": { "type": "integer", "minimum": 1, "maximum": 50 },
          "teamAName": { "type": "string", "maxLength": 80 },
          "teamAId": { "type": "string" },
          "teamAPlayers": { "type": "array", "maxItems": 30, "items": { "$ref": "#/components/schemas/PlayerInput" } },
          "teamBName": { "type": "string", "maxLength": 80 },
          "teamBId": { "type": "string" },
          "teamBPlayers": { "type": "array", "maxItems": 30, "items": { "$ref": "#/components/schemas/PlayerInput" } },
          "venue": { "type": "string", "maxLength": 200 },
          "matchDate": {
            "oneOf": [
              { "type": "string", "format": "date" },
              { "type": "string", "format": "date-time" }
            ]
          },
          "tournamentId": { "type": "string" }
        },
        "description": "Each team requires either its name or a saved team ID. Custom matches also require customOvers.",
        "additionalProperties": true
      },
...
      "Match": {
        "type": "object",
        "properties": {
          "_id": { "type": "string" },
          "matchName": { "type": "string" },
          "matchType": { "type": "string" },
          "venue": { "type": "string" },
          "matchDate": { "type": "string", "format": "date-time" },
          "status": { "type": "string", "enum": ["scheduled", "live", "completed", "cancelled"] },
          "teamA": { "type": "object", "additionalProperties": true },
          "teamB": { "type": "object", "additionalProperties": true }
        },
        "additionalProperties": true
      },
      "MatchResponse": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" },
          "data": { "$ref": "#/components/schemas/Match" }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": { "type": "integer" },
          "limit": { "type": "integer" },
          "total": { "type": "integer" },
          "totalPages": { "type": "integer" },
          "hasNextPage": { "type": "boolean" },
          "hasPreviousPage": { "type": "boolean" }
        }
      },
      "MatchListResponse": {
        "type": "object",
        "required": ["success", "count", "data", "meta"],
        "properties": {
          "success": { "type": "boolean", "example": true },
          "count": { "type": "integer" },
          "data": { "type": "array", "items": { "$ref": "#/components/schemas/Match" } },
          "meta": { "$ref": "#/components/schemas/PaginationMeta" }
        }
      }
    }
  }
}